:truck: Move credentials into api submodule
